Adjusted player hitbox, fixed my sound stuff and adjusted to be an enumerator for falloff distance, fixed gun hold positions - all weapons' origins placed where they seems reasonable to be held

+++ Assets/Scripts/PlayerScripts/PlayerAudio.cs
@@ -22,7 +22,7 @@ public class PlayerAudio : NetworkBehaviour {
 		if (step != null && step.clip != null)
 		{
 			footSource.clip = step.clip;
-			footSource.maxDistance = step.maxDistance;
+			footSource.maxDistance = (int) step.maxDistance;
 			footSource.Play();
             Debug.Log("Played footstep sound!");
 		}
@@ -49,7 +49,7 @@ public class PlayerAudio : NetworkBehaviour {
 		if (landClip != null && landClip.clip != null)
 		{
         	footSource.clip = landClip.clip;
-            footSource.maxDistance = landClip.maxDistance;
+            footSource.maxDistance = (int) landClip.maxDistance;
         	footSource.Play();
         }
         else Debug.Log("Land sound is null!");
@@ -76,7 +76,7 @@ public class PlayerAudio : NetworkBehaviour {
 		if (jumpClip != null && jumpClip.clip != null)
 		{
         	mouthSource.clip = jumpClip.clip;
-            mouthSource.maxDistance = jumpClip.maxDistance;
+            mouthSource.maxDistance = (int) jumpClip.maxDistance;
         	mouthSource.Play();
         }
         else Debug.Log("Jump sound is null!");
@@ -87,7 +87,7 @@ public class PlayerAudio : NetworkBehaviour {
 		if (swapClip != null && swapClip.clip != null)
 		{
 			weaponSource.clip = swapClip.clip;
-            weaponSource.maxDistance = swapClip.maxDistance;
+            weaponSource.maxDistance = (int) swapClip.maxDistance;
 			weaponSource.Play();
         }
         else Debug.Log("Swap sound is null!");
@@ -98,7 +98,7 @@ public class PlayerAudio : NetworkBehaviour {
         if (fireClip != null && fireClip.clip != null)
         {
             weaponSource.clip = fireClip.clip;
-            weaponSource.maxDistance = fireClip.maxDistance;
+            weaponSource.maxDistance = (int) fireClip.maxDistance;
             weaponSource.Play();
             Debug.Log("Played fire sound!");
         }
